+++ crates/icu/src/ic/structures/btreemap.rs
@@ -32,7 +32,7 @@ where
     /// Returns an iterator over all cloned `(K, V)` pairs.
     pub fn iter_pairs(&self) -> impl Iterator<Item = (K, V)> + '_ {
         self.iter()
-            .map(|entry| (entry.key().clone(), entry.value().clone()))
+            .map(|entry| (entry.key().clone(), entry.value()))
     }
 
     /// Returns an iterator over a range of cloned `(K, V)` pairs.
@@ -41,7 +41,7 @@ where
         R: RangeBounds<K>,
     {
         self.range(range)
-            .map(|entry| (entry.key().clone(), entry.value().clone()))
+            .map(|entry| (entry.key().clone(), entry.value()))
     }
 
     /// clear