+++ crates/icu/src/state/mod.rs
@@ -1,3 +1,4 @@
+mod delegated_sessions;
 mod root;
 
 use crate::ic::api::performance_counter;
@@ -6,6 +7,7 @@ use serde::{Deserialize, Serialize};
 use std::cell::RefCell;
 use thiserror::Error as ThisError;
 
+pub use delegated_sessions::{DelegatedSessions, Delegation, DelegationError, RegisterSessionArgs};
 pub use root::{
     CanisterAttributes, CanisterData, CanisterRegistry, CanisterRegistryData, CanisterRegistryError,
 };
@@ -18,6 +20,9 @@ pub use root::{
 pub enum StateError {
     #[error(transparent)]
     CanisterRegistryError(#[from] CanisterRegistryError),
+
+    #[error(transparent)]
+    DelegationError(#[from] DelegationError),
 }
 
 thread_local! {