+++ crates/icu/src/utils/wasm.rs
@@ -1,12 +1,4 @@
-use crate::TC;
 use sha2::{Digest, Sha256};
-
-// format_tc
-#[must_use]
-#[allow(clippy::cast_precision_loss)]
-pub fn format_cycles(cycles: u128) -> String {
-    format!("{:.6} TC", cycles as f64 / TC as f64)
-}
 
 // get_wasm_hash
 #[must_use]